modified receiving PO and purchase return modules

+++ db/migrate/20160312005511_drop_table_product_price_codes.rb
@@ -1,5 +1,5 @@
 class DropTableProductPriceCodes < ActiveRecord::Migration
   def change
-    drop_table :product_price_codes
+    drop_table :product_price_codes if ActiveRecord::Base.connection.table_exists? 'product_price_codes'
   end
 end
